moved the thingy magiggy
